+++ client/public/configs/mock.data.js
@@ -1,3 +1,5 @@
+import {categoryList} from './constants.js'
+
 export const productList = [
   {
     thumbnail: 'https://i.guim.co.uk/img/media/8a13052d4db7dcd508af948e5db7b04598e03190/0_294_5616_3370/master/5616.jpg?width=1200&height=1200&quality=85&auto=format&fit=crop&s=bcaa4eed2c1e6dab61c41a61e41433d9',
@@ -125,4 +127,14 @@ export const product = {
   price: 5600000,
   isLiked: true,
   isOwner: true
+};
+
+export const EDITOR_STATE = {
+  imageList: ['https://i.guim.co.uk/img/media/8a13052d4db7dcd508af948e5db7b04598e03190/0_294_5616_3370/master/5616.jpg?width=1200&height=1200&quality=85&auto=format&fit=crop&s=bcaa4eed2c1e6dab61c41a61e41433d9', 'https://i.guim.co.uk/img/media/8a13052d4db7dcd508af948e5db7b04598e03190/0_294_5616_3370/master/5616.jpg?width=1200&height=1200&quality=85&auto=format&fit=crop&s=bcaa4eed2c1e6dab61c41a61e41433d9', 'https://i.guim.co.uk/img/media/8a13052d4db7dcd508af948e5db7b04598e03190/0_294_5616_3370/master/5616.jpg?width=1200&height=1200&quality=85&auto=format&fit=crop&s=bcaa4eed2c1e6dab61c41a61e41433d9'],
+  name: '빈티지 롤러 스케이트',
+  category: { id: '', title: '' },
+  categoryList: categoryList,
+  price: 169000,
+  contents: '어린시절 추억의 향수를 어쩌구 저쩌구',
+  location: '방이동'
 };
